+++ src/FamilyHubs.SharedKernel.Razor/Dashboard/ColumnImmutable.cs
@@ -1,4 +1,11 @@
 ﻿
 namespace FamilyHubs.SharedKernel.Razor.Dashboard;
 
-public record ColumnImmutable(string DisplayName, string? SortName = null);
+public enum Align
+{
+    Left,
+    Right
+}
+
+// alternatively pass class for header and cell (or both combined) to allow for more customization
+public record ColumnImmutable(string DisplayName, string? SortName = null, Align Align = Align.Left);