+++ docs/ReleaseNotes.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+### 5.8.0
+
+Add Align parameter to ColumnImmutable constructor, to allow the alignment of dashboard columns to be specified.
+
+Remove deprecated `fh-new-tab` attribute from `a` tag helper.
+
 ### 5.7.0
 
 Add support for actions column to summary list row tag helper.